backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.19
boto3==1.28.1
PyJWT==2.7.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0
Flask-SQLAlchemy
Werkzeug